+++ issuer-application/src/main/java/ch/admin/bj/swiyu/issuer/infrastructure/web/signer/WellKnownController.java
@@ -42,6 +42,10 @@ public class WellKnownController {
     private final DemonstratingProofOfPossessionService demonstratingProofOfPossessionService;
     private final MetadataService metadataService;
 
+    private static boolean expectsSignedResponse(String acceptHeader) {
+        return acceptHeader.contains(CONTENT_TYPE_APPLICATION_JWT);
+    }
+
     /**
      * OpenID Connect information for the OAuth Authorization Server
      * This endpoint is a duplicate of the getOpenIDConfiguration endpoint in order to generate a clean openapi doc
@@ -76,7 +80,7 @@ public class WellKnownController {
             return metadataService.getSignedIssuerMetadata(tenantId);
         }
 
-        return metadataService.getUnsignedIssuerMetadata();
+        return metadataService.getUnsignedIssuerMetadata(tenantId);
     }
 
     @GetMapping(value = {"/{tenantId}/.well-known/openid-configuration", "/oid4vci/{tenantId}/.well-known/openid-configuration", "/oid4vci/{tenantId}/.well-known/oauth-authorization-server", "/{tenantId}/.well-known/oauth-authorization-server"})
@@ -89,10 +93,6 @@ public class WellKnownController {
             return metadataService.getSignedOpenIdConfiguration(tenantId);
         }
 
-        return metadataService.getUnsignedOpenIdConfiguration();
-    }
-
-    private static boolean expectsSignedResponse(String acceptHeader) {
-        return acceptHeader.contains(CONTENT_TYPE_APPLICATION_JWT);
+        return metadataService.getUnsignedOpenIdConfiguration(tenantId);
     }
 }